Complete statistics module

+++ src/util/StudentValidator.java
@@ -5,12 +5,14 @@ public class StudentValidator {
 //    ================NAME VALIDATION=============
 
     public static boolean isValidName(String name){
+
         return name!=null && !name.trim().isEmpty();
     }
 
 //    ==============AGE VALIDATION================
 
     public static boolean isValidAge(int age){
+
         return age>=15 && age <=35;
     }
 
@@ -23,18 +25,20 @@ public class StudentValidator {
 //    ===============PHONE VALIDATION===============
 
     public static boolean isValidPhone(String phone){
+
         return phone!=null && phone.matches("\\d{10}");
     }
 
 //    ==================SEMESTER VALIDATION===========
 
     public static boolean isValidSemester(int semester){
+
         return semester >= 1 && semester <=8;
     }
 
 //    ====================CGPA VALIDATION=============
 
-    public static boolean isValidCgpa(int cgpa){
+    public static boolean isValidCgpa(double cgpa){
         return cgpa>=0 && cgpa <=10;
     }
 }
